changed type from int to long for userId

+++ src/main/java/godevenner/divexbackend/portfolio/controller/PortfolioController.java
@@ -39,7 +39,7 @@ public class PortfolioController {
     }
 
     @GetMapping("/portfolioentries")
-    public ResponseEntity<List<PortfolioEntry>> getPortfolioEntries(@RequestBody Integer userId) {
+    public ResponseEntity<List<PortfolioEntry>> getPortfolioEntries(@RequestBody Long userId) {
         List<PortfolioEntry> portfolioentries = portfolioEntryService.getPortfolioEntries(userId);
         return ResponseEntity.ok(portfolioentries);
     }
